1.2.3.test

+++ Assets/Scripts/Controllers/IStats.cs
@@ -1,4 +1,4 @@
 ﻿public interface IStats
 {
-    void ReceiveDamage(float damage);
+    void ReceiveDamage(int damage);
 }
